+++ habapp_rules/core/type_of_day.py
@@ -4,8 +4,6 @@ import datetime
 
 import holidays
 
-from habapp_rules import TIMEZONE
-
 _BY_HOLIDAYS = holidays.country_holidays("DE", "BY")
 
 
@@ -18,7 +16,7 @@ def is_weekend(offset_days: int = 0) -> bool:
     Returns:
         True if desired day is a weekend day
     """
-    day_to_check = datetime.datetime.now(tz=TIMEZONE) + datetime.timedelta(days=offset_days)
+    day_to_check = datetime.datetime.now() + datetime.timedelta(days=offset_days)
     return day_to_check.isoweekday() in {6, 7}
 
 
@@ -31,5 +29,5 @@ def is_holiday(offset_days: int = 0) -> bool:
     Returns:
         True if desired day is holiday
     """
-    day_to_check = datetime.datetime.now(tz=TIMEZONE) + datetime.timedelta(days=offset_days)
+    day_to_check = datetime.datetime.now() + datetime.timedelta(days=offset_days)
     return day_to_check in _BY_HOLIDAYS